backend/requirements.txt
# Django and REST framework
Django==5.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
cx_Oracle==8.3.0

# Authentication
djangorestframework-simplejwt==5.3.1
django-allauth==0.61.1

# API Documentation
drf-yasg==1.21.7

# Development Tools
python-dotenv==1.0.1
django-debug-toolbar==4.3.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Testing
pytest==8.0.2
pytest-django==4.8.0

# Code Quality
black==24.2.0
flake8==7.0.0
isort==5.13.2 